port speech-to-text to docker container
