+++ extension.js
@@ -8,6 +8,7 @@ const spawn = require('child_process')
 // your extension is activated the very first time the command is executed
 
 function activate(context) {
+    var mountDir = __dirname + "/micContainer/containerFiles:/src/micListener";
 
     // Use the console to output diagnostic information (console.log) and errors (console.error)
     // This line of code will only be executed once when your extension is activated
@@ -21,50 +22,10 @@ function activate(context) {
 
         // Display a message box to the user
         vscode.window.showInformationMessage('Start using your microphone!');
-
-        /* 
-        The mic instance is handled as a JSON input, whereby we can set the defaults for the microphone.
-        I had to brute force the device name, because the default 'mic' module did not pick it up.
-        Also, it had channels set to MONO, which my computer does not have as it's default...the default is Stereo (2)
-        */
-
-        /* Info on the micInstance can be found here: https://www.npmjs.com/package/mic */
-        var micInstance = mic({
-            rate: '16000',
-            channels: '2',
-            debug: true,
-            exitOnSilence: 6,
-            fileType: '.wav',
-            device: 'hw:0,0'
-        });
-
-        /*
-        - This is where we start the mic instance, using the above presets.
-        getAudioStream() returns a Transform stream. 
-        - We use a fs stream to create an output file.
-        - The micInputStream is then written to the outputFile, using pipe() from node's Stream API,
-        in order to open in Python.
-        - This file will get rewritten everytime the mic is stopped and started up again.
-        */
-        var micInputStream = micInstance.getAudioStream();
-        const streamDest = __dirname + "/micAudio/output.wav";
-        const pyScriptLoc = __dirname + "/speech.py";
-        var outputFileStream = fs.createWriteStream(streamDest);
-        micInputStream.pipe(outputFileStream);
-
-
-        var startMic = new Promise(function (resolve, reject) {
-            micInstance.start();
-            setTimeout(function () {
-                micInstance.stop();
-                resolve();
-            }, 6000)
-            // outputFileStream.close();
-
-        })
-
-        startMic.then(function () {
-            spawn.exec('python ' + pyScriptLoc, {cwd: __dirname}, (error, stdout, stderr) => {
+        
+        //the way this mounts a mic will only work on linux based machines that have microphone hardware mounted as /dev/snd!!!
+        spawn.exec('docker run -v /dev/snd:/dev/snd -v ' + mountDir + ' --privileged mic-container npm start', (error, stdout, stderr) => {
+            //spawn.exec('docker run -v ./micAudio:/src/micAudio -t argh ')
                 console.log("going to gather the words...")
                 if (error) {
                     console.error(`exec error: ${error}`);
@@ -72,10 +33,9 @@ function activate(context) {
                 }
                 console.log(`Words you spoke: ${stdout}`);
             });
-        })
-
+//old code was here
     });
-
+    
     context.subscriptions.push(disposable);
 }
 exports.activate = activate;
