+++ micContainer/containerFiles/micAudio/placeholder.txt
@@ -1,0 +1,1 @@
+this is where audio recorded from the microphone appears :)
